refactor(frontend): promote matrix primitives to ui
