+++ apps/frontend/src/lib/demos/AdminDashboard.stories.svelte
@@ -13,7 +13,7 @@
   import StatCard from '$lib/components/admin/StatCard.svelte';
   import Panel from '$lib/ui/Panel.svelte';
   import DataTable from '$lib/ui/DataTable.svelte';
-  import CopyId from '$lib/components/admin/CopyId.svelte';
+  import CopyId from '$lib/ui/CopyId.svelte';
   import { Button } from '$lib/ui/form';
   import Pill from '$lib/ui/Pill.svelte';
