+++ apps/frontend/src/lib/ui/CopyId.svelte
