+++ apps/frontend/src/routes/chat/[serverId]/manage/server/members/[userId]/MemberOverviewPanel.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import type { AdminMember, AdminRoleDetails } from '$lib/api-client/adminUsers';
-  import { CopyId } from '$lib/components/admin';
+  import { CopyId } from '$lib/ui';
   import Panel from '$lib/ui/Panel.svelte';
   import UserAvatar from '$lib/components/UserAvatar.svelte';
   import { m } from '$lib/i18n/messages';
